refactor: remove compaction service and integrate context compaction directly into orchestrator
