+++ crates/forge_domain/src/services.rs
@@ -2,9 +2,9 @@ use std::path::Path;
 use std::sync::Arc;
 
 use crate::{
-    Agent, Attachment, ChatCompletionMessage, CompactionResult, Context, Conversation,
-    ConversationId, Environment, File, McpConfig, Model, ModelId, ResultStream, Scope, Tool,
-    ToolCallContext, ToolCallFull, ToolDefinition, ToolName, ToolResult, Workflow,
+    Attachment, ChatCompletionMessage, CompactionResult, Context, Conversation, ConversationId,
+    Environment, File, McpConfig, Model, ModelId, ResultStream, Scope, Tool, ToolCallContext,
+    ToolCallFull, ToolDefinition, ToolName, ToolResult, Workflow,
 };
 
 #[async_trait::async_trait]
@@ -39,11 +39,6 @@ pub trait McpConfigManager: Send + Sync {
 pub trait McpService: Send + Sync {
     async fn list(&self) -> anyhow::Result<Vec<ToolDefinition>>;
     async fn find(&self, name: &ToolName) -> anyhow::Result<Option<Arc<Tool>>>;
-}
-
-#[async_trait::async_trait]
-pub trait CompactionService: Send + Sync {
-    async fn compact_context(&self, agent: &Agent, context: Context) -> anyhow::Result<Context>;
 }
 
 #[async_trait::async_trait]
@@ -128,7 +123,6 @@ pub trait Services: Send + Sync + 'static + Clone {
     type TemplateService: TemplateService;
     type AttachmentService: AttachmentService;
     type EnvironmentService: EnvironmentService;
-    type CompactionService: CompactionService;
     type WorkflowService: WorkflowService;
     type SuggestionService: SuggestionService;
     type McpConfigManager: McpConfigManager;
@@ -139,7 +133,6 @@ pub trait Services: Send + Sync + 'static + Clone {
     fn template_service(&self) -> &Self::TemplateService;
     fn attachment_service(&self) -> &Self::AttachmentService;
     fn environment_service(&self) -> &Self::EnvironmentService;
-    fn compaction_service(&self) -> &Self::CompactionService;
     fn workflow_service(&self) -> &Self::WorkflowService;
     fn suggestion_service(&self) -> &Self::SuggestionService;
     fn mcp_config_manager(&self) -> &Self::McpConfigManager;
